feat(model-number)rename to model number component
